+++ models/pet_store.rb
@@ -49,26 +49,31 @@ class PetStore
   end
 
   def delete
-    begin
-      sql = "DELETE FROM pet_stores WHERE id = #{@id};"
-      SqlRunner.run(sql)
-      @@stores ||= []
-      # self may not be on the @@stores list but a store with the same id should be
-      i = @@stores.find_index { | store | store.id == @id }
-      @@stores.delete_at(i) if i
-      return @id
-    rescue
-      # There's a good chance the delete will fail
-      puts "The delete failed."
-      return nil
+    if @id
+      begin
+        sql = "DELETE FROM pet_stores WHERE id = #{@id};"
+        SqlRunner.run(sql)
+        @@stores ||= []
+        # self may not be on the @@stores list but a store with the same id should be
+        i = @@stores.find_index { | store | store.id == @id }
+        @@stores.delete_at(i) if i
+        @id = nil
+      rescue
+        # There's a good chance the delete will fail
+        puts "The delete failed. Check you have removed all the pets from the store before deleting it."
+      end
     end
+    return @id
   end
 
   def update
-    sql = "UPDATE pet_stores SET (name, address, stock_type) = ('#{@name}', '#{@address}', '#{@stock_type}') WHERE id = #{@id} RETURNING *;"
-    id = SqlRunner.run(sql).first['id']
-    i = @@stores.find_index { | store | store.id == @id }
-    @@stores[i] = self.clone if i
-    return id
+    if @id
+      sql = "UPDATE pet_stores SET (name, address, stock_type) = ('#{@name}', '#{@address}', '#{@stock_type}') WHERE id = #{@id} RETURNING *;"
+      id = SqlRunner.run(sql).first['id']
+      i = @@stores.find_index { | store | store.id == @id }
+      @@stores[i] = self.clone if i && @@stores[i] != self
+      return id
+    end
+    return nil
   end
 end
